+++ packages/natives/native/index.js
@@ -58,6 +58,7 @@ export const matchesKittySequence = nativeBindings.matchesKittySequence;
 export const matchesLegacySequence = nativeBindings.matchesLegacySequence;
 export const parseKey = nativeBindings.parseKey;
 export const parseKittySequence = nativeBindings.parseKittySequence;
+export const processExec = nativeBindings.processExec;
 export const readImageFromClipboard = nativeBindings.readImageFromClipboard;
 export const search = nativeBindings.search;
 export const sliceWithWidth = nativeBindings.sliceWithWidth;